Align per diem table with migrated table conventions
